- reverts FLAVOR_WIN changes from #399 to fix build targeting windows (emulator and the new Windows native build this commit is adding) - Added LONG and ULONG typedefs for compat. with WIn32 builds - Fixed calls to Interlocked API declarations and calls to use LONG to match windows API (Its still a 32 bit integer so no functional change) - updated Lwip and sockets vcxproj files to VS14 - Added DeviceCode\Targets\OS\Win32 for the native Win32 OS porting - Added Solutions\Windows for native Win32 OS port solution - Update solution file to VS14 as it wasn't saved for last commit - reverting #392 as it was choking on malformed paths - updated uvision files, git ignore and added readme.md for crypto

+++ DeviceCode/include/tinyhal.h
@@ -736,13 +736,13 @@ void IDelayLoop2( int iterations );
 //--//
 
 #if !defined(PLATFORM_WINDOWS)
-extern "C" INT32 InterlockedIncrement( volatile INT32* lpAddend );
-extern "C" INT32 InterlockedDecrement( volatile INT32* lpAddend );
-extern "C" INT32 InterlockedExchange( volatile INT32* Target, INT32 Value );
-extern "C" INT32 InterlockedCompareExchange( INT32* Destination, INT32 Exchange, INT32 Comperand );
-extern "C" INT32 InterlockedExchangeAdd( volatile INT32* Addend, INT32 Value );
-extern "C" INT32 InterlockedOr( volatile INT32* Destination, INT32 Flag );
-extern "C" INT32 InterlockedAnd( volatile INT32* Destination, INT32 Flag );
+extern "C" INT32 InterlockedIncrement( volatile LONG* lpAddend );
+extern "C" INT32 InterlockedDecrement( volatile LONG* lpAddend );
+extern "C" INT32 InterlockedExchange( volatile LONG* Target, INT32 Value );
+extern "C" INT32 InterlockedCompareExchange( LONG* Destination, LONG Exchange, LONG Comperand );
+extern "C" INT32 InterlockedExchangeAdd( volatile LONG* Addend, LONG Value );
+extern "C" INT32 InterlockedOr( volatile LONG* Destination, LONG Flag );
+extern "C" INT32 InterlockedAnd( volatile LONG* Destination, LONG Flag );
 #endif // PLATFORM_WINDOWS
 
 struct OpaqueQueueNode
